refactor: split original Zod generator into multiple files to make it more maintainable
